Retry many times when loosing connection to mgr
When asking for a reference to an actor, keep trying to connect to the manager many times. This helps when the application is quitting and the manager finishes before the actors, so exceptions are not propagated on exit.  It also helps if the manager goes away momentarily and then returns, covering for any glitch on the socket/network connection.